+++ drivers/media/platform/tegra/imx091.c
@@ -2838,8 +2838,9 @@ static long imx091_ioctl(struct file *file,
 		return 0;
 
 	case _IOC_NR(NVC_IOCTL_STATIC_RD):
-		dev_dbg(&info->i2c_client->dev, "%s STATIC_RD n=%d\n",
-			__func__, sizeof(struct nvc_imager_static_nvc));
+		dev_dbg(&info->i2c_client->dev, "%s STATIC_RD n=%lu\n",
+			__func__,
+			(unsigned long)sizeof(struct nvc_imager_static_nvc));
 		data_ptr = &info->sdata;
 		if (copy_to_user((void __user *)arg,
 				 data_ptr,
